Downloads/Duphuggingface/requirements.txt
streamlit>=1.50.0
pandas>=2.0.0
beautifulsoup4>=4.14.0
nltk>=3.9.0
plotly>=5.0.0
pytz>=2023.3
streamlit-lightweight-charts>=0.7.0
psycopg2-binary>=2.9.0